Refactoring fetch-based code after PR, again #541, #582 (#723)
As a result of neglecting #582, it became impossible to merge, so I
recreated the Pull-request.
Compared to the last time, we developed it by narrowing down the changes
to make it less likely to cause conflicts.

## Development steps

1. make axios as a optionalDependency
2. move http.ts as http-axios.ts
3. create new http-fetch.ts

as a result, legacy API will depends on the `http-axios.ts`. and so
openapi based clients are depend on `http-fetch.ts`.