Stop using doc branch to use github pages (#868)
`gh-pages` branch for documentation is no longer necessary. By using
GitHub Actions, you can utilize GitHub Pages directly.

This change supports that. Additionally, this change has made sure not
to update the documentation until the upload to npm is complete. This is
to prevent confusion among users if the documentation is updated but
package is not updated.

TODO: Modify settings in
https://github.com/line/line-bot-sdk-nodejs/settings/pages after this
change is merged

ref: https://vitepress.dev/guide/deploy#github-pages

Resolve #521