PR must fail when generated code is not committed (#1046)
When some code are changed by modifying code generator or templates,
forgetting to commit the expected changes can lead to mistakes. I want
to enforce that the PR author commits the generated code, by causing the
CI to fail if the generated code is not committed. This PR achieves
that.