Improve a way to update package version on release (#1057)
The string `__LINE_BOT_SDK_NODEJS_VERSION__` is used in several files to
make it easier to replace with the new version at the time of release.
However, this string is inappropriate as a version, so it needs to be
replaced when testing locally... This is cumbersome for us, so this PR
addresses that issue.

At the time of release, the files that need to be updated will be
directly modified anyway. `./scripts/update-version.mjs` will handle
this. Since 4 lines across 3 files need to be changed, the CI will fail
if this condition is not met. Reviewers can verify that the version is
changed locally, but it's up to them whether to do so.