Bye ts-node (#1112)
ts-node has been present since the beginning of the project, but it is
no longer used. If needed in the future, tsx can be used, and with
Node.js v23, execution as TypeScript will be supported by default
(though it's not perfect yet), so tsx might not even be necessary. For
now, ts-node is no longer needed.