Add EoL and new release reminder (#1189)
This change utilizes https://endoflife.date to create reminders. The
line-bot-sdk repositories use this site as a reference when changing
library version support. After an EOL or a new release is created, if an
issue with the same name has not yet been issued, an issue as a reminder
will be created automatically.

The script (`.github/scripts/check-eol-newrelease.cjs`) itself is the
same across all repositories. The configuration passed from the GitHub
workflow to the script only needs to account for differences specific to
each repository. Only Node.js and Java consider LTS versions, while all
other versions are monitored. It has already been tested in my
repository.