Skip create PR when only git submodule is updated (#1210)
Even when only the submodule is updated and the code is not, a PR is
created in the bot SDK repository.(like
https://github.com/line/line-bot-sdk-nodejs/pull/1208)
If there is no diff in the generated code, I think it's sufficient for
Renovate to update the submodule version without creating a PR through
the GitHub Actions workflow.

Therefore, this change excludes the submodule from the PR creation
conditions and CI failure conditions.