Allow to inject default header from clients (#1230)
This update allows additional headers to be set from the bot SDK's
client. The behavior is as follows:
1. Non-critical headers, such as the user agent, can be overwritten.
2. critical header like authorization headers cannot be overriden.

The goal is to allow the user agent to be overwritten when using this in
other projects. Since it doesn't need to be limited to the user agent
header, it has been implemented more generally.