Reminder for npm audit fix (#1357)
Most npm package alerts can be fixed automatically with `npm audit fix
--force` (it is not perfect, so some manual work is still required).
Because we cannot use a GitHub App token, a reminder should be enough.
Automatically creating PRs would be pointless, because the CI jobs will
not run, and it may be old(=merging may not resolve all issues)

For now, let's create a reminder as an issue. We review issues
regularly, so this should be sufficient.