Set npm version as 11.6.0 in renovate (#1440)
We’re seeing an issue where libraries listed in optionalDependencies are
being removed from the lockfile by Renovate. Specifically, axios is
dropped from the lockfile, causing CI to fail. We found that pinning npm
to 11.6.0 or below resolves the problem, so we’ll update the Renovate
configuration to use npm 11.6.0.

However, I can’t reproduce the issue locally even with npm 11.6.1 or
later, so the root cause hasn’t been determined.